+++ History.md
@@ -6,6 +6,7 @@
 
 *   [#35](https://github.com/pmd/pmd-regression-tester/issues/35): exclude-pattern hasn't been implemented
 *   [#37](https://github.com/pmd/pmd-regression-tester/issues/37): Render stack traces properly
+*   [#40](https://github.com/pmd/pmd-regression-tester/issues/40): NoMethodError on beta3
 
 ## External Contributions
 
@@ -13,6 +14,7 @@
 *   [#34](https://github.com/pmd/pmd-regression-tester/pull/34): Clear old reports before generating new differences regression reports - [BBG](https://github.com/djydewang)
 *   [#36](https://github.com/pmd/pmd-regression-tester/pull/36): Implements exclude-pattern - [BBG](https://github.com/djydewang)
 *   [#39](https://github.com/pmd/pmd-regression-tester/pull/39): Render stacktraces properly - [BBG](https://github.com/djydewang)
+*   [#41](https://github.com/pmd/pmd-regression-tester/pull/41): Fixes NoMethodError in Project class - [BBG](https://github.com/djydewang)
 
 # 1.0.0.beta3 / 2018-08-01